Add files via upload
Public transportation is the main form of transportation utilized in rural Nepalese communities and most people commutes based on posted bus schedules. There is a lack of real-time monitoring of buses which results in commuters waiting for extended periods of time. To address these issues, I conceptualized an idea and have been developing an android based GPS tracking mobile application that will monitor public transportation vehicles in real-time using GPS, Geo-coding, and Google Map API. I developed the Graphical User Interface on Android Studio by utilizing Java and XML.  The goal is to develop a two-view app; a driver-focused view for the bus driver to update the database with their current location and a commuter-focused view for the public to monitor the buses. It is anticipated that this system could be used as a model to address similar issues in other communities that lack advanced technological infrastructure.